Gore Works.
May have broken many parts of the game in testing (I know.)
But the particle system works to emit blood from hit moles with splatters and stamps the ground and splattered particles have a chance to also stamp the ground. uneven ground is going to be a technical challenge to paint, but from testing this...it will be worth it. Hilarious.