Alpha V4 is out
Revamped the menu to fit the theme of the game. Tried to add a loading screen to help with abysmal load times, but it seems like the new menu does not have the same long load times... not sure. Also added third level. No labels are on the main menu, levels are 1-3 from left to right.